Socket refactoring started
