+++ main/CMakeLists.txt
@@ -1,2 +1,2 @@
-idf_component_register(SRCS "tcp_client.c" "wifi.c" "system_simulation.c" "main.c"
+idf_component_register(SRCS "status_socket.c" "wifi.c" "system_simulation.c" "main.c" "utils.c"
                        INCLUDE_DIRS ".")
